backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
requests
beautifulsoup4
selenium
undetected-chromedriver
webdriver-manager
python-dotenv
supabase
firecrawl-py
schedule
pytz
openai
instructor
setuptools
requests-html
anthropic
boto3
cloudscraper
google-cloud-texttospeech
pandas
urllib3
Pillow